requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
PyJWT==2.8.0

Pillow==10.0.0
EOF